Remove MediaService.getMimeType()
Forcing the _mimeType to x-media-{servicename} ties my hands a bit.
The initial content item is created and then a read-only version is
saved with _mimeType=x-media-kaltura. The UX would look for
kaltura/video, kaltura/audio, or kaltura/image. So nothing renders.

This way version 1 of the content will have the correct _mimeType
and the KalturaMesiaService is free to handle multiple kinds of
media.